% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortNoteExtract.R
\name{cohortNoteExtract}
\alias{cohortNoteExtract}
\title{cohortNoteExtract function}
\usage{
cohortNoteExtract(num, date, result = FALSE, resultdb, cohort, noteType)
}
\arguments{
\item{num}{The number of clinical reports want to extract}

\item{date}{The note dates of the reports. It should contain the min date and max date.}

\item{result}{If TRUE, then the Note of the Cohort population will be extracted. Default is FALSE}

\item{resultdb}{The database which contains the cohort result of the ATLAS}

\item{cohort}{The cohort id generated from ATLAS}

\item{noteType}{The note type concept ids of the NOTE table. It uses the standardized concepts of OMOP CDM}
}
\description{
cohortNoteExtract function extracts NOTE data from the Cohort generated fron ATLAS, an open source data analysis platform of OMOP CDM.
}
