% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseConnection.R
\name{databaseConnection}
\alias{databaseConnection}
\title{databaseConnection function}
\usage{
databaseConnection(server, user, password, schema)
}
\arguments{
\item{server}{server information of the database}

\item{user}{user id of the database}

\item{password}{password of the database}

\item{schema}{schema of the database want to access}
}
\description{
databaseConnection function is using DatabaseConnector package. The createConnectionDetails and connect functions are congested into a single function.
}
