% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{dictionary}
\alias{dictionary}
\title{Dictionary function}
\usage{
dictionary(x, y)
}
\arguments{
\item{x}{an expression want to be replaced or removed}

\item{y}{an expression want to replace}
}
\description{
Dictionary function
}
