% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawNotedateGraph.R
\name{drawNotedateGraph}
\alias{drawNotedateGraph}
\title{drawNotedateGraph function}
\usage{
drawNotedateGraph(Text)
}
\arguments{
\item{Text}{text data from Preprocessing step}
}
\description{
drawNotedateGraph function
}
