% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawOptimalTopicNumber.R
\name{drawOptimalTopicNumber}
\alias{drawOptimalTopicNumber}
\title{databaseConnection function}
\usage{
drawOptimalTopicNumber(dtm, TopicNum, By, metrics)
}
\arguments{
\item{dtm}{DocumentTermMatrix want to explore.}

\item{TopicNum}{The range of topic numbers. It should contain at least to numbers}

\item{By}{The interval of topics when searching}

\item{metrics}{String or vector of possible metrics: "Griffiths2004", "CaoJuan2009", "Arun2010", "Deveaud2014".}
}
\description{
databaseConnection function is using DatabaseConnector package. The createConnectionDetails and connect functions are congested into a single function.
}
