% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawPersonTable.R
\name{drawPersonTable}
\alias{drawPersonTable}
\title{drawPersonTable function}
\usage{
drawPersonTable(Text)
}
\arguments{
\item{Text}{text data from Preprocessing step}
}
\description{
drawPersonTable function
}
