% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonEditor.R
\name{jsonEditor}
\alias{jsonEditor}
\title{jsonEditor}
\usage{
jsonEditor(filePath)
}
\arguments{
\item{filePath}{The file path which indicates the location of the JSON file when uploading it.}
}
\description{
jsonEditor is a JSON viewer using listviewer. JsonEditor enables to upload or download the JSON files and easily can edit.
}
