% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonTemplateEditor.R
\name{jsonTemplateEditor}
\alias{jsonTemplateEditor}
\title{jsonTemplateEditor}
\usage{
jsonTemplateEditor(filePath)
}
\arguments{
\item{filePath}{The file path which indicates the location of the JSON file when uploading it.}
}
\description{
jsonTemplateEditor is a JSON viewer using listviewer package.
}
