% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcess.R
\name{preProcess}
\alias{preProcess}
\title{preProcess}
\usage{
preProcess(text, english = F, whitespace = F, stopwords = F,
  number = F, punc = F, stem = F, lower = F)
}
\arguments{
\item{text}{text corpus want to analyze.}

\item{english}{if TRUE then alphabets except English will be deleted from the courpus.}

\item{whitespace}{if TRUE then extra whitespaces will be deleted from the courpus.}

\item{stopwords}{if TRUE then only English stopwords will be deleted. The stopwords are from tm package.}

\item{number}{if TRUE then numbers will be deleted from the corpus}

\item{punc}{if TRUE then punctuations will be deleted from the courpus.}

\item{stem}{if TRUE then terms in the corpus will be stemmed. The stemming logic is from tm package.}

\item{lower}{if TRUE then English characters will be lowered. If other languages except English are included the function will not work.}
}
\description{
This is a preprocessing function which can perform basic proprocessing rules.
}
