% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveJson.R
\name{saveJson}
\alias{saveJson}
\title{saveJson}
\usage{
saveJson(JSON)
}
\arguments{
\item{JSON}{Previously defined empty object to conatin JSON data}
}
\description{
This function save previously annotated data using SOCRATex into JSON files.
}
