% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyLDA.R
\name{shinyLDA}
\alias{shinyLDA}
\title{shinyLDA}
\usage{
shinyLDA(dtm, k, iter, alpha)
}
\arguments{
\item{dtm}{DocumentTerm Matrix which can be used for LDA analysis.}

\item{k}{The number of topics(clusters) want to analyze.}

\item{iter}{The number of Learning iterations.}

\item{alpha}{Tha alpha parameter of the Latent Dirichlet Allocation.}
}
\description{
This function is implementing LDAvis into RShint form especially for SOCRATex application.
}
