% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicSample.R
\name{topicSample}
\alias{topicSample}
\title{topicSample}
\usage{
topicSample(Text, topic, num)
}
\arguments{
\item{Text}{The text data want to analyze}

\item{topic}{The number of topics in LDA analysis}

\item{num}{The number of sample documents want to observe}
}
\description{
This function shows the sample documents of the LDA clustering result. To implement the function LDAvis analysis need to be performed.
}
