% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnigramTokenizer.R
\name{UnigramTokenizer}
\alias{UnigramTokenizer}
\title{UnigranTokenizer function}
\usage{
UnigramTokenizer(x)
}
\arguments{
\item{x}{text data which can be tokenized}
}
\description{
This is a uni-gram tokenizer for creating Document-Term Matrix
}
