% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useDictionary.R
\name{useDictionary}
\alias{useDictionary}
\title{useDictionary}
\usage{
useDictionary(useDictionary = FALSE)
}
\arguments{
\item{useDictionary}{If it is TRUE then it reads the dictionary database and perform preprocessing}
}
\description{
This function preprocess data based on the user customized pre-defined dictionary. Currerntly, it used the dictionary in Sql Server.
The dictionary needs to has two columns, first is the exporession
}
