% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unigramTokenizer.R
\name{unigramTokenizer}
\alias{unigramTokenizer}
\title{unigranTokenizer function}
\usage{
unigramTokenizer(x)
}
\arguments{
\item{x}{text data which can be tokenized}
}
\description{
This is a uni-gram tokenizer for creating Document-Term Matrix
}
